Dedup lineage-edit engine into one shared module (server + WASM worker)
The merge / edit-lineage-root / undo (conflict-aware) / lineage-hierarchy logic was
duplicated byte-for-byte between the Express backend (server.js) and the in-browser
local backend worker (lineageEditing.js). Extract it into a single shared CommonJS
module, src/ui/shared/lineageEditCore.cjs, used by both, so the two modes can't drift.

- server.js: the 6 editing endpoints (/lineages, /merge-lineage, /edit-lineage-root,
  /undo-edit, /undo-preview, /edit-history) delegate to the shared core; removed the
  inline editHistory + rebuildCladeLabels / snapshotTips / computeConflictingEdits
  (and the dead restoreTipSnapshot).
- lineageEditing.js: delegates to the shared core; keeps only the genuinely
  client-specific bits (the export 'modified' baseline + jsonl/metadata/pb builders).
- dev.sh: mount src/ui/shared into the container (the Dockerfile already COPYs it).

originalLineages stays per-caller: the server captures all nodes with a multi-field
fallback on load, the worker captures tips lazily — and the pb export legitimately
differs — so those aren't shared.

Verified: node --check; backendless vite build bundles the CJS core into the ESM
worker; live server run drives merge -> parent-merge -> undo(0) with the conflict
cascade [0,1] removing both, an idempotent round-trip, and edit-root validation.
