+++ src/ui/taxonium_backend/server.js
@@ -130,7 +130,11 @@ waitForTheImports = async () => {
 var processedData = null;
 var cached_starting_values = null;
 var originalLineages = new Map();
-var editHistory = [];  // Stack of edit operations for undo
+
+// Lineage-edit engine (merge / edit-root / undo / conflict / hierarchy) — the SAME
+// code the in-browser worker runs, so the two modes can't drift. It owns editHistory
+// (module-global, one per server process, as before).
+const editCore = require("../shared/lineageEditCore.cjs");
 
 function snapshotLineages() {
   originalLineages.clear();
@@ -618,137 +622,8 @@ app.get("/config", function (req, res) {
 });
 
 app.get("/lineages", function (req, res) {
-  const start_time = Date.now();
-  console.log("/lineages - extracting all lineage data from full dataset");
-  console.log("DEBUG: /lineages endpoint called");
-
-  // Get the field name from query parameter, default to 'meta_annotation_1'
   const field = req.query.field || 'meta_annotation_1';
-  console.log("DEBUG: field =", field);
-
-  // Build node lookup and parent-child relationships for tree traversal FIRST
-  const nodeLookup = {};
-  const children = {};
-  let totalNodes = 0;
-  let nodesWithLineage = 0;
-
-  processedData.nodes.forEach(node => {
-    totalNodes++;
-    nodeLookup[node.node_id] = node;
-    if (node[field] && node[field] !== '') {
-      nodesWithLineage++;
-    }
-    if (node.parent_id && node.parent_id !== node.node_id) {
-      if (!children[node.parent_id]) {
-        children[node.parent_id] = [];
-      }
-      children[node.parent_id].push(node.node_id);
-    }
-  });
-
-  console.log(`DEBUG: Built parent-child relationships for ${Object.keys(children).length} parents`);
-
-  // Function to get all descendant node IDs recursively
-  const getAllDescendants = (nodeId) => {
-    const descendants = [];
-    const toVisit = children[nodeId] || [];
-
-    while (toVisit.length > 0) {
-      const currentNodeId = toVisit.pop();
-      descendants.push(currentNodeId);
-
-      // Add children of current node to visit queue
-      if (children[currentNodeId]) {
-        toVisit.push(...children[currentNodeId]);
-      }
-    }
-
-    return descendants;
-  };
-
-  // Build lineage hierarchy directly from clade labels in the tree
-  const cladeNodeMap = new Map(); // Map clade label -> clade node
-  const lineageHierarchy = new Map(); // Map lineage -> {count, descendants, parent, etc.}
-
-  // First pass: collect all clade nodes
-  for (const node of processedData.nodes) {
-    if (node.clades && node.clades.pango && !node.is_tip) {
-      cladeNodeMap.set(node.clades.pango, node);
-    }
-  }
-
-  // Helper function to find parent lineage by walking up the tree
-  // Returns the first ancestor clade node with a different label
-  const findParentLineage = (nodeId, currentCladeLabel) => {
-    let currentNode = nodeLookup[nodeId];
-    while (currentNode && currentNode.parent_id && currentNode.parent_id !== currentNode.node_id) {
-      const parentNode = nodeLookup[currentNode.parent_id];
-      if (parentNode && parentNode.clades && parentNode.clades.pango) {
-        const parentClade = parentNode.clades.pango;
-        if (parentClade !== currentCladeLabel) {
-          return parentClade;
-        }
-      }
-      currentNode = parentNode;
-    }
-    return null;
-  };
-
-  // Second pass: for each clade node, calculate counts and find parent from tree
-  // Include ALL clade nodes (even those without direct tips — they may be
-  // intermediate lineages needed for a correct hierarchy)
-  for (const [cladeLabel, cladeNode] of cladeNodeMap.entries()) {
-    const descendants = getAllDescendants(cladeNode.node_id);
-
-    const descendantLineageSet = new Set();
-    let totalTips = 0;
-    let directTips = 0;
-
-    descendants.forEach(descId => {
-      const descNode = nodeLookup[descId];
-      if (descNode && descNode.is_tip) {
-        totalTips++;
-        if (descNode[field] && descNode[field] !== '') {
-          if (descNode[field] !== cladeLabel) {
-            descendantLineageSet.add(descNode[field]);
-          } else {
-            directTips++;
-          }
-        }
-      }
-    });
-
-    const parentLineage = findParentLineage(cladeNode.node_id, cladeLabel);
-
-    lineageHierarchy.set(cladeLabel, {
-      value: cladeLabel,
-      count: totalTips,
-      descendantLineages: descendantLineageSet.size,
-      descendantLeaves: totalTips,
-      parent: parentLineage
-    });
-  }
-
-  console.log(`DEBUG: Found ${totalNodes} total nodes`);
-  console.log(`DEBUG: Found ${cladeNodeMap.size} clade nodes`);
-  console.log(`DEBUG: Built hierarchy for ${lineageHierarchy.size} lineages`);
-
-  // Convert to array format from our clade-based hierarchy
-  const lineageArray = Array.from(lineageHierarchy.values());
-
-  // Sort by count descending
-  lineageArray.sort((a, b) => b.count - a.count);
-
-  const response = {
-    lineages: lineageArray,
-    field: field,
-    totalNodes: totalNodes,
-    nodesWithLineage: nodesWithLineage,
-    uniqueLineages: lineageArray.length
-  };
-
-  console.log(`Found ${lineageArray.length} unique lineages in ${Date.now() - start_time}ms`);
-  res.send(response);
+  res.send(editCore.getLineages(processedData, field));
 });
 
 app.get("/mutations/", function (req, res) {
@@ -854,478 +729,41 @@ app.get("/nodes/", function (req, res) {
 
 // POST endpoint to merge lineages - combines child lineages into parent
 app.post("/merge-lineage", function (req, res) {
-  const start_time = Date.now();
-  console.log("/merge-lineage - merging lineage assignments");
-
   const { lineageName, field } = req.body;
-
-  if (!lineageName || !field) {
-    return res.status(400).send({
-      error: "Missing required parameters: lineageName, field"
-    });
-  }
-
-  // Determine the parent lineage from the tree structure
-  // Find the clade node for this lineage, then walk up to find the parent clade
-  let parentLineage = null;
-  const nodeLookup = {};
-  processedData.nodes.forEach(node => { nodeLookup[node.node_id] = node; });
-
-  // Find this lineage's clade node
-  let cladeNode = null;
-  for (const node of processedData.nodes) {
-    if (!node.is_tip && node.clades && node.clades.pango === lineageName) {
-      cladeNode = node;
-      break;
-    }
-  }
-
-  if (cladeNode) {
-    // Walk up tree to find parent clade (including the root node)
-    let current = nodeLookup[cladeNode.parent_id];
-    while (current) {
-      if (current.clades && current.clades.pango && current.clades.pango !== lineageName) {
-        parentLineage = current.clades.pango;
-        break;
-      }
-      if (current.parent_id === current.node_id) break; // reached root
-      current = nodeLookup[current.parent_id];
-    }
-  }
-
-  if (!parentLineage) {
-    return res.status(400).send({
-      error: `Cannot merge "${lineageName}" - no parent lineage found in tree`
-    });
-  }
-
-  console.log(`Merging ${lineageName} (and sub-lineages) into ${parentLineage}`);
-
-  // Merge: reassign tips with this lineage or any sub-lineage to the parent
-  const isLineageToMerge = (nodeLineage) => {
-    return nodeLineage === lineageName || nodeLineage.startsWith(lineageName + '.');
-  };
-
-  // Snapshot affected tips before the edit
-  const affectedNodeIds = new Set();
-  processedData.nodes.forEach(node => {
-    if (node.is_tip && node[field] && isLineageToMerge(node[field])) {
-      affectedNodeIds.add(node.node_id);
-    }
-  });
-  const snapshot = snapshotTips(field, affectedNodeIds);
-
-  let mergedCount = 0;
-  const affectedLineages = new Set();
-
-  processedData.nodes.forEach(node => {
-    if (node.is_tip && node[field] && isLineageToMerge(node[field])) {
-      affectedLineages.add(node[field]);
-      node[field] = parentLineage;
-      mergedCount++;
-    }
-  });
-
-  // Rebuild clade labels to reflect the new tip assignments
-  rebuildCladeLabels(field);
-
-  // Record in edit history
-  const editEntry = {
-    id: editHistory.length,
-    action: 'merge',
-    lineageName,
-    parentLineage,
-    field,
-    description: `Merged ${lineageName} into ${parentLineage} (${mergedCount} nodes)`,
-    timestamp: new Date().toISOString(),
-    snapshot,
-    affectedLineages: [lineageName, parentLineage, ...affectedLineages],
-  };
-  editHistory.push(editEntry);
-
-  console.log(`Merged ${mergedCount} nodes into ${parentLineage} in ${Date.now() - start_time}ms`);
-
-  res.send({
-    success: true,
-    mergedCount,
-    parentLineage,
-    affectedLineages: Array.from(affectedLineages),
-    message: `Merged ${mergedCount} samples into ${parentLineage}`
-  });
+  const result = editCore.mergeLineage(processedData, lineageName, field);
+  if (result.error) return res.status(400).send({ error: result.error });
+  res.send(result);
 });
-
-// Rebuild clade labels by finding the MRCA of each lineage's tips
-function rebuildCladeLabels(field) {
-  console.log('Rebuilding clade labels...');
-  const start = Date.now();
-
-  const nodeLookup = {};
-  processedData.nodes.forEach(node => { nodeLookup[node.node_id] = node; });
-
-  // Clear existing clade labels
-  processedData.nodes.forEach(node => {
-    if (!node.is_tip && node.clades) {
-      delete node.clades.pango;
-    }
-  });
-
-  // Collect tips per lineage
-  const lineageTips = {};
-  processedData.nodes.forEach(node => {
-    if (node.is_tip && node[field]) {
-      const lin = node[field];
-      if (!lineageTips[lin]) lineageTips[lin] = [];
-      lineageTips[lin].push(node.node_id);
-    }
-  });
-
-  // Compute depth for each node (used to find deepest common ancestor)
-  const depth = {};
-  processedData.nodes.forEach(node => {
-    if (depth[node.node_id] !== undefined) return;
-    const stack = [];
-    let cur = node;
-    while (depth[cur.node_id] === undefined) {
-      stack.push(cur.node_id);
-      if (cur.parent_id === cur.node_id) { depth[cur.node_id] = 0; stack.pop(); break; }
-      cur = nodeLookup[cur.parent_id];
-    }
-    let d = depth[cur.node_id];
-    while (stack.length > 0) {
-      d++;
-      depth[stack.pop()] = d;
-    }
-  });
-
-  // LCA of two nodes: walk the deeper one up, then walk both up together
-  const lca = (a, b) => {
-    let na = a, nb = b;
-    while (depth[na] > depth[nb]) na = nodeLookup[na].parent_id;
-    while (depth[nb] > depth[na]) nb = nodeLookup[nb].parent_id;
-    while (na !== nb) {
-      na = nodeLookup[na].parent_id;
-      nb = nodeLookup[nb].parent_id;
-    }
-    return na;
-  };
-
-  // For each lineage, find MRCA and set clade label
-  let cladeCount = 0;
-  for (const [lineage, tips] of Object.entries(lineageTips)) {
-    if (tips.length === 0) continue;
-    let mrca = tips[0];
-    for (let i = 1; i < tips.length; i++) {
-      mrca = lca(mrca, tips[i]);
-    }
-    const mrcaNode = nodeLookup[mrca];
-    if (mrcaNode && !mrcaNode.is_tip) {
-      if (!mrcaNode.clades) mrcaNode.clades = {};
-      mrcaNode.clades.pango = lineage;
-      cladeCount++;
-    }
-  }
-
-  console.log(`Rebuilt ${cladeCount} clade labels in ${Date.now() - start}ms`);
-}
-
-// Snapshot current tip assignments for a set of node IDs (for undo)
-function snapshotTips(field, nodeIds) {
-  const snapshot = {};
-  for (const node of processedData.nodes) {
-    if (node.is_tip && nodeIds.has(node.node_id)) {
-      snapshot[node.node_id] = node[field] || '';
-    }
-  }
-  return snapshot;
-}
-
-// Restore tip assignments from a snapshot and rebuild clade labels
-function restoreTipSnapshot(field, snapshot) {
-  for (const node of processedData.nodes) {
-    if (node.is_tip && snapshot.hasOwnProperty(node.node_id)) {
-      node[field] = snapshot[node.node_id];
-    }
-  }
-  rebuildCladeLabels(field);
-}
 
 // GET endpoint to retrieve edit history
 app.get("/edit-history", function (req, res) {
-  res.send(editHistory.map(entry => ({
-    id: entry.id,
-    action: entry.action,
-    lineageName: entry.lineageName,
-    parentLineage: entry.parentLineage,
-    description: entry.description,
-    timestamp: entry.timestamp,
-    affectedLineages: entry.affectedLineages || [],
-  })));
+  res.send(editCore.getEditHistory());
 });
-
-// Compute which edits would need to be undone along with a target edit.
-// An edit conflicts if it modified any of the same node IDs as the target
-// or any other edit already in the conflict set (transitive closure).
-function computeConflictingEdits(targetId) {
-  const targetIndex = editHistory.findIndex(e => e.id === targetId);
-  if (targetIndex === -1) return [];
-
-  const target = editHistory[targetIndex];
-  const taintedNodeIds = new Set(Object.keys(target.snapshot));
-  const toUndo = [targetId];
-
-  // Walk forward from target+1, expanding the tainted set transitively
-  for (let i = targetIndex + 1; i < editHistory.length; i++) {
-    const entry = editHistory[i];
-    const entryNodeIds = Object.keys(entry.snapshot);
-    const conflicts = entryNodeIds.some(id => taintedNodeIds.has(id));
-    if (conflicts) {
-      toUndo.push(entry.id);
-      entryNodeIds.forEach(id => taintedNodeIds.add(id));
-    }
-  }
-  return toUndo;
-}
 
 // GET endpoint to preview which edits would be undone for a given edit id
 app.get("/undo-preview/:id", function (req, res) {
-  const targetId = parseInt(req.params.id, 10);
-  const ids = computeConflictingEdits(targetId);
-  res.send({ targetId, wouldUndo: ids });
+  res.send(editCore.undoPreview(parseInt(req.params.id, 10)));
 });
 
 // POST endpoint to undo an edit and any conflicting later edits
 app.post("/undo-edit", function (req, res) {
-  const { id } = req.body || {};
-
-  if (editHistory.length === 0) {
-    return res.status(400).send({ error: "No edits to undo" });
+  const result = editCore.undoEdit(processedData, (req.body || {}).id);
+  if (result.error) {
+    const code = /not found/.test(result.error) ? 404 : 400;
+    return res.status(code).send({ error: result.error });
   }
-
-  // Default to last edit
-  const targetId = id !== undefined ? id : editHistory[editHistory.length - 1].id;
-  const targetIndex = editHistory.findIndex(e => e.id === targetId);
-  if (targetIndex === -1) {
-    return res.status(404).send({ error: `Edit ${targetId} not found` });
-  }
-
-  const idsToUndo = new Set(computeConflictingEdits(targetId));
-
-  // Collect entries to undo (in history order) and entries to keep
-  const toUndo = [];
-  const toKeep = [];
-  for (const entry of editHistory) {
-    if (idsToUndo.has(entry.id)) {
-      toUndo.push(entry);
-    } else {
-      toKeep.push(entry);
-    }
-  }
-
-  // Restore snapshots in reverse order
-  for (let i = toUndo.length - 1; i >= 0; i--) {
-    const entry = toUndo[i];
-    console.log(`Undoing edit #${entry.id}: ${entry.description}`);
-    for (const node of processedData.nodes) {
-      if (node.is_tip && entry.snapshot.hasOwnProperty(node.node_id)) {
-        node[entry.field] = entry.snapshot[node.node_id];
-      }
-    }
-  }
-
-  // Rebuild clade labels once
-  rebuildCladeLabels(toUndo[0].field);
-
-  // Replace history with only the kept entries (re-index ids)
-  editHistory.length = 0;
-  toKeep.forEach((entry, i) => {
-    entry.id = i;
-    editHistory.push(entry);
-  });
-
-  console.log(`Undid ${toUndo.length} edit(s), kept ${toKeep.length}`);
-
-  res.send({
-    success: true,
-    undone: toUndo[0].description,
-    removedCount: toUndo.length,
-    removedIds: toUndo.map(e => e.id),
-    remainingEdits: editHistory.length,
-  });
+  res.send(result);
 });
 
 // POST endpoint to edit lineage root assignments based on selected tree node
 app.post("/edit-lineage-root", function (req, res) {
-  const start_time = Date.now();
-  console.log("/edit-lineage-root - reassigning lineage based on tree structure");
-
   const { lineageName, rootNodeId, field } = req.body;
-
-  if (!lineageName || !rootNodeId || !field) {
-    return res.status(400).send({
-      error: "Missing required parameters: lineageName, rootNodeId, field"
-    });
+  const result = editCore.editLineageRoot(processedData, lineageName, rootNodeId, field);
+  if (result.error) {
+    const code = /not found/.test(result.error) ? 404 : 400;
+    return res.status(code).send({ error: result.error });
   }
-
-  console.log(`Reassigning ${lineageName} to start from node ${rootNodeId} for field ${field}`);
-
-  // Find the root node (handle both string and number IDs)
-  const rootNodeIdNum = parseInt(rootNodeId, 10);
-  const rootNode = processedData.nodes.find(node =>
-    node.node_id === rootNodeId || node.node_id === rootNodeIdNum
-  );
-  if (!rootNode) {
-    return res.status(404).send({
-      error: `Root node ${rootNodeId} not found`
-    });
-  }
-
-  // Build a lookup for faster parent/child relationships
-  const nodeLookup = {};
-  const children = {};
-  processedData.nodes.forEach(node => {
-    nodeLookup[node.node_id] = node;
-    if (node.parent_id && node.parent_id !== node.node_id) {
-      if (!children[node.parent_id]) {
-        children[node.parent_id] = [];
-      }
-      children[node.parent_id].push(node.node_id);
-    }
-  });
-
-  // Function to get all descendant nodes of a given node
-  const getDescendants = (nodeId) => {
-    const descendants = new Set();
-    const queue = [nodeId];
-
-    while (queue.length > 0) {
-      const currentId = queue.shift();
-      descendants.add(currentId);
-
-      if (children[currentId]) {
-        children[currentId].forEach(childId => {
-          if (!descendants.has(childId)) {
-            queue.push(childId);
-          }
-        });
-      }
-    }
-
-    return descendants;
-  };
-
-  // Build a set of all lineages that are children of the edited lineage
-  // (so we don't overwrite more-specific annotations)
-  const childLineages = new Set();
-  processedData.nodes.forEach(node => {
-    if (!node.is_tip && node.clades && node.clades.pango) {
-      const clade = node.clades.pango;
-      if (clade !== lineageName && clade.startsWith(lineageName + '.')) {
-        childLineages.add(clade);
-      }
-      // Also check auto. children: auto.lineageName.X is a child
-      if (clade.startsWith('auto.' + lineageName + '.')) {
-        childLineages.add(clade);
-      }
-    }
-  });
-  // Also collect child lineages from tip annotations
-  processedData.nodes.forEach(node => {
-    if (node.is_tip && node[field]) {
-      const ann = node[field];
-      if (ann !== lineageName && ann.startsWith(lineageName + '.')) {
-        childLineages.add(ann);
-      }
-      if (ann.startsWith('auto.' + lineageName + '.')) {
-        childLineages.add(ann);
-      }
-    }
-  });
-
-  const isChildLineage = (ann) => childLineages.has(ann);
-
-  // Get all nodes that should have this lineage (descendants of root node)
-  const targetNodeIds = getDescendants(rootNodeId);
-
-  // Find the parent lineage to reassign displaced tips
-  let parentLineage = null;
-  let cur = nodeLookup[rootNode.parent_id];
-  while (cur) {
-    if (cur.clades && cur.clades.pango && cur.clades.pango !== lineageName) {
-      parentLineage = cur.clades.pango;
-      break;
-    }
-    if (cur.parent_id === cur.node_id) break; // reached root
-    cur = nodeLookup[cur.parent_id];
-  }
-
-  // Snapshot all tips that could be affected (in subtree + currently assigned outside)
-  const affectedNodeIds = new Set();
-  processedData.nodes.forEach(node => {
-    if (!node.is_tip) return;
-    if (targetNodeIds.has(node.node_id)) affectedNodeIds.add(node.node_id);
-    if ((node[field] || null) === lineageName) affectedNodeIds.add(node.node_id);
-  });
-  const snapshot = snapshotTips(field, affectedNodeIds);
-
-  let assignedCount = 0;
-  let clearedCount = 0;
-
-  processedData.nodes.forEach(node => {
-    if (!node.is_tip) return;
-    const currentLineage = node[field] || null;
-
-    if (targetNodeIds.has(node.node_id)) {
-      // Inside the new subtree: assign this lineage UNLESS the tip has a
-      // more-specific child lineage annotation
-      if (currentLineage !== lineageName && !isChildLineage(currentLineage)) {
-        node[field] = lineageName;
-        assignedCount++;
-      }
-    } else if (currentLineage === lineageName) {
-      // Outside the new subtree: reassign to parent lineage
-      if (parentLineage) {
-        node[field] = parentLineage;
-      } else {
-        node[field] = '';
-      }
-      clearedCount++;
-    }
-  });
-
-  console.log(`Assigned ${lineageName} to ${assignedCount} tips, displaced ${clearedCount} tips`);
-
-  rebuildCladeLabels(field);
-
-  // Collect all lineages this edit touched
-  const editAffectedLineages = new Set([lineageName]);
-  if (parentLineage) editAffectedLineages.add(parentLineage);
-  for (const val of Object.values(snapshot)) {
-    if (val) editAffectedLineages.add(val);
-  }
-
-  // Record in edit history
-  const editEntry = {
-    id: editHistory.length,
-    action: 'edit-root',
-    lineageName,
-    field,
-    description: `Moved ${lineageName} root to node ${rootNodeId} (${assignedCount} assigned, ${clearedCount} displaced)`,
-    timestamp: new Date().toISOString(),
-    snapshot,
-    affectedLineages: [...editAffectedLineages],
-  };
-  editHistory.push(editEntry);
-
-  console.log(`Operation completed in ${Date.now() - start_time}ms`);
-
-  res.send({
-    success: true,
-    assignedCount,
-    clearedCount,
-    totalAffected: targetNodeIds.size,
-    parentLineage,
-    message: `Reassigned ${lineageName} to ${assignedCount} nodes under root ${rootNodeId}`
-  });
+  res.send(result);
 });
 
 function startListening() {
